+++ spec/strategy/archive_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe EY::Serverside::Strategy::Archive do
+describe EY::Serverside::Source::Archive do
   before do
     described_class.any_instance.stub(:runner) { RunnerDouble }
   end
 
-  context "strategy" do
+  context "source" do
     let(:shell) { ShellDouble.new }
     subject {
       described_class.new(shell,